Rename AnnotationHelpers to FunctionParserHelper
