+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/annotations/ScalarImplementationHeader.java
@@ -25,7 +25,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static com.facebook.presto.metadata.FunctionRegistry.mangleOperatorName;
-import static com.facebook.presto.operator.annotations.AnnotationHelpers.parseDescription;
+import static com.facebook.presto.operator.annotations.FunctionsParserHelper.parseDescription;
 import static com.google.common.base.CaseFormat.LOWER_CAMEL;
 import static com.google.common.base.CaseFormat.LOWER_UNDERSCORE;
 import static com.google.common.base.Preconditions.checkArgument;
